# perf: cache block data for CreditPool for asset unlock limits calculation