## Issue being fixed or feature implemented
For each new block CreditPool read blocks twice from disk: for last connected block and for distant block (576 block ago) which is used to calculate sliding window limit for withdrawals.

## What was done?
Added mini cache with block data (locked amount, indexes) to avoid reading block with ReadBlockFromDisk.
[It is possible](https://github.com/knst/dash/tree/perf-cp-cache-2) to avoid 2nd block reading too (for tip), but benchmark doesn't show clear improvement. It will go to the separate PR later.

## How Has This Been Tested?
develop:
<img width="613" alt="image" src="https://github.com/user-attachments/assets/b29382cf-2de3-4223-a85e-2623982ff86a" />
```
2025-05-04T18:36:29Z [bench]       - ProcessSpecialTxsInBlock: 52.94ms [84.66s (14.46ms/blk)]
2025-05-04T18:36:29Z [bench]       - CheckCreditPoolDiffForBlock: 0.21ms [2.11s (0.36ms/blk)]
2025-05-04T18:36:29Z [bench]   - Connect total: 54.40ms [94.48s (16.14ms/blk)]
```

RP:
<img width="613" alt="image" src="https://github.com/user-attachments/assets/40166b85-cc37-4bf3-a618-2931c3d0fdca" />
```
2025-05-05T10:11:47Z [bench]       - ProcessSpecialTxsInBlock: 52.62ms [83.01s (14.18ms/blk)]
2025-05-05T10:11:47Z [bench]       - CheckCreditPoolDiffForBlock: 0.21ms [2.09s (0.36ms/blk)]
2025-05-05T10:11:47Z [bench]   - Connect total: 53.46ms [90.66s (15.49ms/blk)]
```

## Breaking Changes
N/A


## Checklist:
- [x] I have performed a self-review of my own code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have added or updated relevant unit/integration/functional/e2e tests
- [ ] I have made corresponding changes to the documentation
- [x] I have assigned this pull request to a milestone